Move root logger fixture out to utils module
