+++ tests/fixtures/utils.py
@@ -15,9 +15,11 @@
 
 """Utils for fixture handling"""
 
+import logging
 import os
 from pathlib import Path
 
+import pytest
 import yaml
 
 TEST_FILE_DIR = Path(__file__).parent.resolve() / "test_files"
@@ -33,3 +35,20 @@ def read_yaml(path: Path) -> dict:
     """Read yaml file and return content as dict."""
     with open(path, encoding="UTF-8") as file:
         return yaml.safe_load(file)
+
+
+@pytest.fixture
+def root_logger_reset():
+    """Reset root logger level and handlers after modification."""
+    root = logging.getLogger()
+    original_level = root.level
+    root_handlers = root.handlers.copy()
+
+    yield
+
+    # reset level and remove RecordCompiler handler
+    root.setLevel(original_level)
+
+    for handler in root.handlers:
+        if handler not in root_handlers:
+            root.addHandler(handler)
